% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_records.R
\name{flatten_records}
\alias{flatten_records}
\title{Flatten User Records}
\usage{
flatten_records(x)
}
\arguments{
\item{x}{A list of dataframes.}
}
\value{
A single dataframe with unnested 'records' and 'value' columns.
}
\description{
This function takes a list, converts each element into a tibble, binds them all into a single dataframe,
and then unnests the 'records' and 'value' columns.
}
