% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_api}
\alias{get_api}
\title{Get the API key for a DID and app password}
\usage{
get_api(fed, did, app_password)
}
\arguments{
\item{fed}{The \href{https://atproto.com}{ATProto} instance to which the handle belongs.}

\item{did}{The DID for which to get the API key.}

\item{app_password}{The password to include in the request.}
}
\value{
The API key corresponding to the provided DID and password.
}
\description{
This function sends a request to the specified URL, includes the provided DID and password in the
request body, performs the request, and returns the API key found in the response body.
}
