% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_did}
\alias{get_did}
\title{Get the DID (Decentralized Identifier) for a handle}
\usage{
get_did(fed = "bsky.social", handle)
}
\arguments{
\item{fed}{The \href{https://atproto.com}{ATProto} instance to which the handle belongs.}

\item{handle}{The handle for which to get the DID.}
}
\value{
The DID corresponding to the provided handle.
}
\description{
This function sends a request to the specified URL, appends the provided handle to the URL path,
performs the request, and returns the DID found in the response body.
}
