% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_records.R
\name{get_user_records}
\alias{get_user_records}
\title{Get User Records}
\usage{
get_user_records(
  request,
  key,
  repo,
  collection,
  count,
  throttle = c(3000, 300)
)
}
\arguments{
\item{request}{The request URL.}

\item{key}{The API key.}

\item{repo}{The repository.}

\item{collection}{The collection.}

\item{count}{The number of records to fetch. If \code{count} is greater than 100, the function will send multiple requests in batches of 100.}

\item{throttle}{A vector specifying the rate limit as \code{c(requests, seconds)}. If the number of requests is divisible by \code{requests}, the function will pause for \code{seconds} seconds. Defaults to \code{c(3000, 300)}.}
}
\value{
A list of responses. Each response is a list representing the JSON response from each request.
}
\description{
This function sends a request to a specified URL, appending necessary details in the URL path,
and fetches user records based on the count parameter. If count is more than 100, additional requests
are sent out in batches of 100 until the count is fulfilled.
}
