% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSky-package.R
\docType{package}
\name{rSky-package}
\alias{rSky}
\alias{rSky-package}
\title{rSky: An R package for interacting with Bluesky.social}
\description{
The goal of rSky is to provide R users with some basic tools for authenticating to and fetching data from Bluesky. It currently has limited functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/scarnecchia/rSky}
  \item \url{https://scarnecchia.github.io/rSky/}
  \item Report bugs at \url{https://github.com/scarnecchia/rSky/issues}
}

}
\author{
\strong{Maintainer}: D. Scarnecchia \email{scarnecchia@gmail.com}

}
\keyword{internal}
